Replaced some uint64_t with uintptr_t

+++ source/Memory.cpp
@@ -4,23 +4,23 @@
 
 std::vector<bool> Memory::g_frames;
 
-uint64_t Memory::getFreePage()
+uintptr_t Memory::getFreePage()
 {
   fDeg() << "free page";
 
-  for (uint64_t i = 0; i < g_frames.size(); ++i)
+  for (uintptr_t i = 0; i < g_frames.size(); ++i)
     if (!g_frames[i])
     {
       g_frames[i] = true;
       return i;
     }
-  uint64_t i = g_frames.size();
+  uintptr_t i = g_frames.size();
   g_frames.resize(i+16, false);
   g_frames[i] = true;
   return i;
 }
 
-void Memory::setPageFree(uint64_t page)
+void Memory::setPageFree(uintptr_t page)
 {
   //if (g_frames.size() <= page)
   //  g_frames.resize(intAlignSup(page+1, 16));
@@ -31,7 +31,7 @@ void Memory::setPageFree(uint64_t page)
   g_frames[page] = false;
 }
 
-void Memory::setPageUsed(uint64_t page)
+void Memory::setPageUsed(uintptr_t page)
 {
   if (g_frames.size() <= page)
     g_frames.resize(intAlignSup(page+1, 16));
